Modulo 7 - Clase 7

+++ project/project/routers/reviews.py
@@ -1,24 +1,23 @@
-from fastapi import HTTPException, APIRouter
+from fastapi import HTTPException, APIRouter, Depends
 
 from typing import List
 
 from ..database import UserReview, User, Movie
 from ..schemas import ReviewRequestModel, ReviewResponseModel, ReviewRequestPutModel
 
+from ..common import get_current_user
+
 
 router = APIRouter(prefix='/reviews')
 
 @router.post('', response_model=ReviewResponseModel)
-async def create_review(user_review: ReviewRequestModel):
-
-    if User.select().where(User.id == user_review.user_id).exists():
-        raise HTTPException(404, 'El usuario no existe.')
+async def create_review(user_review: ReviewRequestModel, user: User = Depends(get_current_user)): 
     
     if Movie.select().where(Movie.id == user_review.movie_id).exists():
         raise HTTPException(404, 'La película no existe.')
     
     user_review = UserReview.create(
-        user_id=user_review.user_id,
+        user_id=user.id,
         movie_id=user_review.movie_id,
         review=user_review.review,
         score=user_review.score 
@@ -44,11 +43,14 @@ async def get_review(review_id: int):
 
 
 @router.put('/{review_id}', response_model=ReviewResponseModel)
-async def update_review(review_id: int, review_request: ReviewRequestPutModel):
+async def update_review(review_id: int, review_request: ReviewRequestPutModel, user: User = Depends(get_current_user)):
     user_review = UserReview.select().where(UserReview.id == review_id).first()
 
     if user_review is None:
         raise HTTPException(404, 'La reseña no existe.')
+    
+    if user_review.user_id != user.id:
+        raise HTTPException(401, 'No eres el propietario.')
     
     user_review.review = review_request.review
     user_review.score = review_request.score
@@ -58,11 +60,14 @@ async def update_review(review_id: int, review_request: ReviewRequestPutModel):
     return user_review
 
 @router.delete('/{review_id}', response_model=ReviewResponseModel)
-async def delete_review(review_id: int):
+async def delete_review(review_id: int, user: User = Depends(get_current_user)):
     user_review = UserReview.select().where(UserReview.id == review_id).first()
 
     if user_review is None:
         raise HTTPException(404, 'La reseña no existe.')
+    
+    if user_review.user_id != user.id:
+        raise HTTPException(401, 'No eres el propietario.')
     
     user_review.delete_instance()
